requirements.txt
fastapi
uvicorn[standard]
pydantic
llama-cpp-python
duckduckgo-search
